Turns out, the way I was getting note values was screwing up certain notes
I’m guessing that some notes had ‘//u’ in the note value? not sure.
Stole the JSON.parse idea from this guy
https://www.npmjs.com/package/soundfont-player

It also turns out that when you use JSON instead of a bunch of arrays,
the notes don’t end up out of order, so the sorting is no longer
necessary. I’m leaving the sorting code in place for now (in a method
called ‘sortNotes’), just in case users need to sort notes at some
point.