Large internal refactor
Created “Sound” class. A “plain sound” is a sound, and a “Note” extends
sound.

Did this to facilitate computed properties on a per-sound basis. CPs
such as “track.currentPosition” and “track.length” should be possible
now.

The only bit of the API that changed is that the promise from “load”
now returns the sound that was loaded, and “pan()” now needs to take
place in that promise, as “pan()” is now called on the note object
itself, so the note object must exist prior to panning.